+++ test/unit/rx-query.test.ts
@@ -17,6 +17,7 @@ import * as humansCollection from './../helper/humans-collection';
 import * as schemaObjects from '../helper/schema-objects';
 import * as schemas from './../helper/schemas';
 import * as util from '../../dist/lib/util';
+import {RxJsonSchema} from '../../src/types';
 
 config.parallel('rx-query.test.js', () => {
     describe('mquery', () => {
@@ -711,16 +712,16 @@ config.parallel('rx-query.test.js', () => {
                 db.destroy();
             });
             it('schema example 2', async () => {
-                const schema = {
+                const schema: RxJsonSchema = {
                     keyCompression: false,
                     version: 0,
                     type: 'object',
                     properties: {
                         value: {
-                            type: 'number',
-                            index: true
+                            type: 'number'
                         }
-                    }
+                    },
+                    indexes: ['value']
                 };
                 const db = await createRxDatabase({
                     name: util.randomCouchString(10),
@@ -757,10 +758,10 @@ config.parallel('rx-query.test.js', () => {
                     version: 0,
                     properties: {
                         name: {
-                            type: 'string',
-                            index: true
+                            type: 'string'
                         }
-                    }
+                    },
+                    indexes: ['name']
                 }
             });
 
@@ -1017,12 +1018,12 @@ config.parallel('rx-query.test.js', () => {
                         type: 'object',
                         properties: {
                             title: {
-                                type: 'string',
-                                index: true
+                                type: 'string'
                             },
                         },
                     }
-                }
+                },
+                indexes: ['info.title']
             };
             const db = await createRxDatabase({
                 name: util.randomCouchString(10),
@@ -1080,7 +1081,7 @@ config.parallel('rx-query.test.js', () => {
             db.destroy();
         });
         it('#609 default index on _id when better possible', async () => {
-            const mySchema = {
+            const mySchema: RxJsonSchema = {
                 version: 0,
                 keyCompression: false,
                 type: 'object',
@@ -1089,10 +1090,10 @@ config.parallel('rx-query.test.js', () => {
                         type: 'string'
                     },
                     passportId: {
-                        type: 'string',
-                        index: true
+                        type: 'string'
                     }
-                }
+                },
+                indexes: ['passportId']
             };
             const collection = await humansCollection.createBySchema(mySchema);
 
@@ -1120,7 +1121,7 @@ config.parallel('rx-query.test.js', () => {
             collection.database.destroy();
         });
         it('#698 Same query producing a different result', async () => {
-            const mySchema = {
+            const mySchema: RxJsonSchema = {
                 version: 0,
                 keyCompression: false,
                 type: 'object',
@@ -1132,10 +1133,10 @@ config.parallel('rx-query.test.js', () => {
                         type: 'string'
                     },
                     created_at: {
-                        type: 'number',
-                        index: true
+                        type: 'number'
                     }
-                }
+                },
+                indexes: ['created_at']
             };
             const collection = await humansCollection.createBySchema(mySchema);
 
@@ -1156,6 +1157,15 @@ config.parallel('rx-query.test.js', () => {
                 $and: [{
                     event_id: {
                         $eq: 2
+                    }
+                }, {
+                    user_id: {
+                        $eq: '6'
+                    }
+                },
+                {
+                    created_at: {
+                        $gt: null
                     }
                 }, {
                     user_id: {